Tidy up; Move files into new folders
